made buttons delete after interaction over

+++ src/buttons/buttons.py
@@ -3,7 +3,7 @@ from discord.ext import commands
 from discord import app_commands
 import sys
 
-from discord.interactions import Interaction
+from discord.interactions import Interaction, InteractionMessage, InteractionResponse
 project_root = 'C:\\Users\\Vinea\\Desktop\\Personal Projects\\ybot'
 sys.path.append(project_root)
 from src.utils.utility import Utility
@@ -13,12 +13,16 @@ class SteamButton(discord.ui.View):
     def __init__(self, interaction: discord.Interaction):
         super().__init__(timeout=None)
         self.interaction = interaction
-        self.cmd_interaction_user = interaction.user
         self.member_id = interaction.user.id
-        print("initialized")
+        self.is_pressed = False
     
     async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if interaction.user != self.interaction.user:
+        if self.is_pressed == True:
+            print("Buttons are disabled!")
+            embed = discord.Embed(color=discord.Color.yellow(), title=f"Choice already made")
+            await interaction.response.send_message(embed=embed, ephemeral=True)
+
+        elif interaction.user != self.interaction.user:
             embed = discord.Embed(color=discord.Color.yellow(), title=f"Try using your own /{self.interaction.command.name}")
             await interaction.response.send_message(embed=embed, ephemeral=True)
             return False
@@ -26,11 +30,7 @@ class SteamButton(discord.ui.View):
     
     @discord.ui.button(label="confirm", style=discord.ButtonStyle.green)
     async def confirm(self, interaction: discord.Interaction, Button: discord.ui.Button):
-        print("debug1")
-        print(self.interaction.user.id)
-        
-        if interaction.user.id == self.member_id:
-            Button.disabled = True
+        if interaction.user.id == self.member_id and self.is_pressed==False:
             if u.unregister(self.interaction.user.id):
                 title = "Confirmed unlinking..!"
                 embedColor = discord.Color.green()
@@ -42,15 +42,18 @@ class SteamButton(discord.ui.View):
             embed = discord.Embed(color=embedColor, title=title)
             await interaction.response.send_message(embed=embed, ephemeral=True)
             if interaction.response.is_done():
+                self.is_pressed = True
                 await interaction.message.delete()
+                
     
     @discord.ui.button(label="cancel", style=discord.ButtonStyle.red)
     async def cancel(self, interaction: discord.Interaction, Button: discord.ui.Button):
         embed = discord.Embed(color=discord.Color.red(), title="Canceled unlinking..!")
-        if interaction.user.id == self.member_id:
-            Button.disabled = True
+        if interaction.user.id == self.member_id and self.is_pressed==False:
             await interaction.response.send_message(embed=embed, ephemeral=True)
             if interaction.response.is_done():
+                self.is_pressed = True
                 await interaction.message.delete()
+
